agent/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp>=1.18.0",
    "openai>=2.6.0",
    "python-dotenv>=1.1.1",
]
